[spark] Simplify catalog-managed Format Table partition DDL
- Treat metastore.partitioned-table as a strict persisted property for a
  Format Table: a conflicting dynamic Spark option now fails the table load
  (via FormatTable#copy) instead of being silently ignored with a warning.
- Move the Format Table partition DDL (add/drop/registration lookup) off the
  shared PaimonPartitionManagement trait onto PaimonFormatTable, dropping the
  nullable partitionManager hook every BaseTable inherited and the reverse
  FormatTable casts; restore the trait's FileStoreTable gate on toPaimonPartitions.
- Collapse the DROP PARTITION rewrite back to a single UnresolvedPaimonRelation
  rule, extended to recognize catalog-managed Format Tables, so the parser no
  longer duplicates the extractor or depends on the execution package.
