+++ paimon-spark/paimon-spark-common/src/main/scala/org/apache/paimon/spark/execution/PaimonStrategy.scala
@@ -176,8 +176,7 @@ case class PaimonStrategy(spark: SparkSession)
     case RepairTable(
           r @ ResolvedTable(_, _, table: PaimonFormatTable, _),
           enableAddPartitions,
-          enableDropPartitions)
-        if PaimonFormatTablePartitionDdlExec.usesCatalogManagedPartitions(table) =>
+          enableDropPartitions) if table.hasCatalogManagedPartitions =>
       PaimonRepairFormatTablePartitionsExec(
         table,
         enableAddPartitions,
