+++ paimon-spark/paimon-spark-common/src/main/scala/org/apache/spark/sql/catalyst/parser/extensions/RewriteSparkDDLCommands.scala
@@ -18,18 +18,12 @@
 
 package org.apache.spark.sql.catalyst.parser.extensions
 
-import org.apache.paimon.spark.SparkTable
 import org.apache.paimon.spark.catalyst.plans.logical.PaimonDropPartitions
-import org.apache.paimon.spark.execution.PaimonFormatTablePartitionDdlExec
-import org.apache.paimon.spark.format.PaimonFormatTable
 
 import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.catalyst.analysis.{EliminateSubqueryAliases, UnresolvedTable}
 import org.apache.spark.sql.catalyst.plans.logical.{DropPartitions, LogicalPlan}
 import org.apache.spark.sql.catalyst.rules.Rule
-import org.apache.spark.sql.connector.catalog.{CatalogManager, LookupCatalog, TableCatalog}
-
-import scala.util.Try
+import org.apache.spark.sql.connector.catalog.{CatalogManager, LookupCatalog}
 
 case class RewriteSparkDDLCommands(spark: SparkSession)
   extends Rule[LogicalPlan]
@@ -42,40 +36,11 @@ case class RewriteSparkDDLCommands(spark: SparkSession)
     // A new member was added to CreatePaimonView since spark4.0,
     // unapply pattern matching is not used here to ensure compatibility across multiple spark versions.
     //
-    // Both Paimon data tables and Format Tables with catalog-managed partitions route DROP
-    // PARTITION through PaimonDropPartitions (the latter resolve partial specs through the
-    // partition gateway, so they must bypass Spark's strict partition-spec resolution too). A
-    // single extractor loads the table once and classifies it, avoiding a second
-    // catalog.loadTable per analyzer iteration. Format Tables using filesystem partition
-    // discovery are not matched and keep Spark's own resolution plus the explicit unsupported
-    // error at execution.
-    case DropPartitions(UnresolvedPaimonDropTarget(aliasedTable), parts, ifExists, purge) =>
+    // Rewrite before Spark resolves the partition specs so both Paimon data tables and Format
+    // Tables with catalog-managed partitions can handle partial DROP specs themselves. Format
+    // Tables using filesystem partition discovery are not matched (see UnresolvedPaimonRelation)
+    // and keep Spark's own resolution plus the explicit unsupported error at execution.
+    case DropPartitions(UnresolvedPaimonRelation(aliasedTable), parts, ifExists, purge) =>
       PaimonDropPartitions(aliasedTable, parts, ifExists, purge)
-  }
-
-  private object UnresolvedPaimonDropTarget {
-    def unapply(plan: LogicalPlan): Option[LogicalPlan] = {
-      EliminateSubqueryAliases(plan) match {
-        case UnresolvedTable(multipartIdentifier, _, _)
-            if isPaimonDropTarget(multipartIdentifier) =>
-          Some(plan)
-        case _ => None
-      }
-    }
-  }
-
-  private def isPaimonDropTarget(multipartIdentifier: Seq[String]): Boolean = {
-    multipartIdentifier match {
-      case CatalogAndIdentifier(catalog: TableCatalog, ident) =>
-        Try(catalog.loadTable(ident))
-          .map {
-            case _: SparkTable => true
-            case formatTable: PaimonFormatTable =>
-              PaimonFormatTablePartitionDdlExec.usesCatalogManagedPartitions(formatTable)
-            case _ => false
-          }
-          .getOrElse(false)
-      case _ => false
-    }
   }
 }
